Extend SigningIdentity with config update signing functions
FAB-17741
